requirements.txt
amqp==5.1.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
blis==0.7.9
catalogue==2.0.8
celery==5.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.3
confection==0.0.4
configobj==5.0.8
configparser==5.3.0
confluent-kafka==2.2.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.5.0/es_core_news_sm-3.5.0-py3-none-any.whl#sha256=13b909bac0de5f5338f4327b14a9cc9736cf51f07131a3cfcb688dbeb8608bdd
et-xmlfile==1.1.0
etelemetry==0.3.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.8.0
fpdf==1.7.2
fqdn==1.5.1
future==0.18.3
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kafka-python==2.0.2
kombu==5.3.2
langcodes==3.3.0
looseversion==1.0.3
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
murmurhash==1.0.9
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nibabel==5.0.0
nipype==1.8.5
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.0
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pathy==0.10.1
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.3.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
prov==2.0.0
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pydot==1.4.2
Pygments==2.15.1
PyJWT==2.5.0
PyMuPDF==1.21.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
pytest==7.4.2
python-barcode==0.14.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.4
pyxnat==1.5
PyYAML==6.0
pyzbar==0.1.9
pyzmq==25.0.2
qrcode==7.3.1
qtconsole==5.4.3
QtPy==2.3.1
rdflib==6.2.0
redis==4.6.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scipy==1.10.0
Send2Trash==1.8.2
sentry-sdk==1.30.0
simplejson==3.18.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
stack-data==0.6.2
terminado==0.17.1
thinc==8.1.9
tinycss2==1.2.1
toml==0.10.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
traits==6.3.2
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.14
vine==5.0.0
virtualenv==20.16.6
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
